Start of where-provenance extraction
For now, we take a traced result and annotate where-provenance
everywhere (at the leaves/base types). Selectively adding provenance
only in parts seems great if one wanted to actually use this, but to
figure out whether getting where-provenance out of a trace works at
all seems unneccessary.

We need traces to be typed, similar to expressions. I suspect we will
also need a trace environment, which sounds problematic if traces need
to be databaseable.

This also restricts tables to be lists of records with values of base
type.