Ugh. What is my actual problem?
When finding the trace of a for comprehension output, we end up with a
trace with free variables.

There must be a way to avoid this. We can't just inline the "value" of
the variable, because it's a trace, like a row in table -- it doesn't
make sense.
However, we could replace variables by their where-provenance. Isn't
that what everyWhere does? Ugh, I'm so confused!