Variants in the object language first try
This works, but not great, because Idris can't prove eval total
anymore. This is because we use Idris functions in Case', which makes
it not strictly positive. There's Expr G a on the left hand side of an
arrow type.

One alternative is to use object language variables and environments
instead. This makes code a bit less pretty, but that's probably fine.