Trace FOR polymorphically
I think this (almost?) makes sense now. The case for if-then-else is a
bit hairy, because we need to refer to the TRACE type function. Doing
it directly breaks polymorphism (we fix a to String). I think we need
to do the same as we do for the value trace analysis function, which
we pass in polymorphically.