Tracing projection
To trace a projection, I need to trace the record expression. To trace
the record expression, I need to provide the type of the record
expression. This is why the type of the record expression is annotated
in the projection expression. This kind of sucks. I see two better
options:

1. Have all expressions annotated with *their* type, so we don't need
to pass the type to the trace function. This would avoid "guessing"
the type of subterms in some other cases too. E.g. the subterms of
list concatenation have the same type. This would be nice, except that
I seem to be unable to annotate variables without giving up the
Constructor Monad instance that saves me from writing my own
instantiate/abstract.

2. Make up a an approximation of the record expression's type. Given a
projection M.l : T(C), we know that M : T(Record (l: C, ...)), or in
words: if we're projecting out field l with type C, we know that the
record contains the field l with type C, and nothing else really
matters. We have a proper way of having holes in records: row
variables. However, making up a fresh row variable is hard. If we make
the record case type directed (that is only trace lables mentioned in
the type) we could just pass a singleton record type and just not
trace the other fields. They are not going to get used anyways.