Now with working distribution over subtraces
This is a bit of a mix of my two styles of self-tracing right-hand
sides. This version avoids producing functions everywhere and
immediately applying them, except for the distribution case. I'm
pretty sure I could avoid that too, or go back to more beta redexes
for the other traces, but it seems to work, so let's stick with the
weird mix for now.