Annotate bound variables with the types recorded in their binders
This makes the tests pass, but I could have changed the generator to
produce annotated variables. Also, this duplicates type
information. I need type information at the variables, but do I really
need it at the binders? I guess I kind of do, to have a type even if
the variable is not used, or to have something to check against if the
same variable is used twice with different type annotations.

Oh and the other thing: maybe I'm using the bound library all
wrong. Maybe if I require my variables to have types, I should require
an argument of type `Expr c a (Type a)` instead of `Expr c a x`. Like
I do in prettyExpr (which should probably be Doc, but nevermind that).
But then, could I use `hoistScope`? I'm confused.