Generalize binary Concat to n-ary and add some commuting conversions
The n-ary Concat is a bit dodgy, because the n=0 case is allowed by
the type, but not properly handled. In theory Empty == Concat [] and
I'd quite like to do that, but Empty takes a type annotation and
Concat does not. Concat does not need one unless it is 0-ary. If I had
a proper type inferencer I wouldn't need the annotation on Empty. I
should probably write one.

isOneNF tests for absence of type and function applications,
tracecase, typecase, and some aspects of the query normal
form (i.e. only iterating over tables). With this I found some missing
commuting conversions.